Break out hash scrolling into it's own file
